website-accessibility-fix/requirements.txt
# Core dependencies for website accessibility fix tool
requests>=2.31.0
psutil>=5.9.0
python-nmap>=0.7.1
dnspython>=2.4.0
paramiko>=3.3.0
pyyaml>=6.0.1
click>=8.1.0
colorama>=0.4.6
tabulate>=0.9.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.82.0

# Development dependencies
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0